editquality: add the WIKI_HOST env variable to model.py
The change adds awareness for the WIKI_HOST environment variable
to model.py, allowing it to set as the HTTP Host header for
the extractor function. This is really handy when the WIKI_URL
is an internal endpoint (like https://api-ro.discovery.wmnet) and not
an internet routable endpoint (like https://en.wikipedia.org).

Bug: T289778
Change-Id: Ib9b234cbe315b68e620df7245638fc4b59b529da