draftquality: null change to trigger image publishing
The last code review to bump draftquality's revscoring to 2.11.4
ended up nowhere since CI Zuul was broken when merging,
so the change didn't get build to a Docker image and
then published to our Docker registry. This change is meant
to be a null op, that will trigger a new image to be built.

Change-Id: Ib5a06c70595955d808eae0d8094a98c04ac9cf5a