editquality: add support for revision-score events
This change adds the first basic support for the generation
of revision-score events. The functionality is already present
in ORES, where the /precache URI expects a revision-create
event JSON and it returns a revision-score event.

The idea is to have Lift Wing to emit events directly
to EventGate, without the need of ChangeProp in the middle
(or other tools like it).

The idea is to do the following:
- Create a new EventGate Stream called mediawiki.revision-score-test.
- Use Airflow or something similar to listen to revision-create
  events, and contact the Lift Wing API with a special parameter.
- Check the kafka topic mediawiki.revision-score-test to see if
  events are correctly validated/generated etc..

Bug: T301878
Change-Id: Ia7d83b3b9a379079d17d93f2ede385a301b35c01