Update Python model servers and requirements to KServe 0.8
KServe 0.8 introduces a rename of classes that we inherit from:
https://github.com/kserve/kserve/releases/tag/v0.8.0

It also requires a new version of revscoring, since the current
one conflicts with the numpy version required by KServe 0.8.

The outlink's requirements.txt needed a downgrade of numpy
due to a conflict between kserve and what we previously
requested, shouldn't be a problem but let me know in case.
The aioredis dependency needs also to be downgraded
for the same reason.

Bug: T311982
Change-Id: I2da98d5d7c42c9ea7d6f79943feffb19c76bf6d0