articlequality: move preprocess() to async
This change has been tested on stat1004 with a basic revscoring
script, verifying that no urllib logs/http-calls were made
with the http cache set.

There seems to be only one API call made from the extractor
to the MW API, so probably the performance gain in this case
is minimal. The code should be able to handle better/more
requests in parallel though.

Bug: T313915
Change-Id: I77aca6477e19c23a092a5807aa2e57b3aa5cfa48