outlink: add code to send events to EventGate
This change enables the API to accept a revision-create event,
generate a revision-score event including predicted topics for
the revision, and send it to EventGate.

Also, add a new environ variable MODEL_VERSION, move get_outlinks
function to the class functions, and add get_lang and get_page_title
in the shared python module.

Bug: T315994
Change-Id: Ic6e768e8cc72f397486bf4812f4485c10ce52220