editquality: refactor preprocess common code
While adding async preprocess support to drafttopic
and draftquality I noticed that a lot of code is simply
copied/pasted, and that we have two common use cases for
the extractor:
- one MW API call to fetch the rev-id's info
  (articlequality and drafttopic).
- three MW API calls to fetch the rev-id's info,
  together with its parent-rev-id and user (editquality
  and draftquality).
The main issue in adding another function to the python
common modules is that it would add the revscoring
dependency to the requirements.txt file as well, and
we'll end up with non-revscoring models in need to add
the revscoring dep only to access common code.
To avoid it, I added a new subdir under `python` called
`revscoring`, with its own requirements.txt file, to separate
concerns and ease the deprecation of revscoring when
the time will come.

Bug: T313915
Change-Id: I0a3fd3e5addb81f90b6efab0a797f2c9f80707a4