events.py: improve revision score generation code
The revision-create schema states that only some fields
are mandatory, meanwhile others are optional. I spotted the issue
while testing Lift Wing with Benthos, since some revision-create
events led to KeyError exceptions.

Moreover Aiko discovered that we are not following the revision-score
schema, since the "scores" field is not added. This change
modifies the output format of the event computed by the model
server according to what revision-score suggests to use in
its examples.

A more generic revision-create validator function could also be
useful, to double check that the event carried by the request
is well formed, but it may become stale very soon. The validation
would in fact be relevant for a specific version of a schema,
so more thoughts need to be put in figuring out a better solution
in my opinion. For the moment we can keep going with this code
to spot more corner cases before we find a definitive solution.

Change-Id: Iee680f9b19289deb6ebf6ddfe7fc01acc1dc4c6a