extractor_utils.py: handle badrevids errors from the MW API
The MW API returns a 'badrevids' error when the revision-id
provided in input is not available/present for a specific
wiki. This seems to happen when the related page has been deleted,
and we currently handle the use case returning an HTTP 500.
(so we return the same error for no revision info and a network
availability issue).

It seems more appropriate to handle the use case with a HTTP 400,
so in the future we'll not have alarms related to these kind of
events.

While at it, break/reformat the error msg returned in
the HTTP response for the more generic feature extraction
error case.

Bug: T320374
Change-Id: I6ed94c536639411d17a38420226be04eeeeda97f