extractor_utils: add fetch_features function
Most of the revscoring model servers (except articlequality)
use the same function to retrieve feature values from
the revscoring extractor. This change adds the function to
extractor_utils.py with some add-ons:
- Catch the more generic MissingResource exception (broader
  range of errors from the extractor are handled).
- Type hints
The function is then applied to the editquality model server,
as a first test.

The black utility also reports an extra line needed to
the preprocess_utils.py. add it as well.

Bug: T320374
Change-Id: I4c01b788eced5ad7cc698924a717b47851a887c3