editquality: improve aiohttp and asyncio performances
In the asyncio documentation it is suggested to avoid
creating sessions for every request (or every few
requests). In our case this means that we can't use
the context manager anymore, and that we should handle
creation/clean-up of the session by ourselves.

Moreover, the asyncio's default executor in KServe
uses a ThreadPool, and its size is auto-calculated
if no value is provided (on our pods it is 5).
Add the possibility to configure the number of
threads used by asyncio.

Bug: T320374
Change-Id: I5d6ef340377e778ecfe7842f8b988b47c74d8db5