Refactor revscoring model servers
After the last refactoring that Aiko did for articlequality's
model server, all the revscoring ones have basically the same
code (except for a few nits). After working with the codebase
I realized that it is very error prone to keep using separate
model-servers for the revscoring use case, so I have introduced
a new class to factor out the common code.

Important notes:
* I think that for the time being we could avoid to have
  a separate requirements.txt for the revscoring use case.
  It is easier to simply track dependencies where needed, to
  avoid another dependency hell back and forth. If we find
  ourselves struggling for it in the future there will be
  the time to refactor it properly.
* I fixed a few bits of the MultiProcessing code, and added
  a way to save cpu cycles when extended_features is turned
  on. It sadly doesn't work with MP yet, I left a note about
  the problem, but since it is related to revscoring/pickling
  I'd suggest to work on it only if stricly necessary.
* The fetch_features function can be heavily cpu-bound,
  it depends on the rev-id from my tests, and adding MP code
  improves a lot the performances.

Bug: T320374
Change-Id: I30727608fea642728fd142f2d4ee56afc5241cae