revertrisk: change output and remove HTTPError exception
We saw a lot of HTTPError of missing information required for inference
when testing batch prediction, but not sure why it happended. To better
understanding the issue, we logged the responses returned by the MW API
in knowledge_integrity.
https://gitlab.wikimedia.org/repos/research/knowledge_integrity/-/commit/beee3dc64e4d453045ea2bd24cd18cf34528ebd4

Various changes:
* Changed the knowledge_integrity tag in requirements.txt to a branch
so we can test changes made to KI.
* Removed the HTTPError exception raised when get_current_revision
returns an empty result in model.py.
* Added the lang and rev_id requested by users to the output, and
returned an empty score if missing revision from preprocess.

Bug: T323023
Change-Id: I1581bec50d8728f7cc57c108596d793a90731e50