Create different objects for MP/SP
Create different objects for multiprocessing and allow to
enable MP only for scoring/inference. The purpose of this
refactoring is to avoid having many if/else statements
in the code and increase maintainability.
We have added the following boolean environment variables: PREPROCESS_MP, INFERENCE_MP
which enable/disable multiprocessing for the corresponding model server part.

Bug: T323624
Change-Id: Id728810eb1f34733151bea8e240efaf2896ed76d