Add pre-commit hooks
Adding checks in file .pre-commit-config.yaml. By default CI will fail once
these changes are included as some files will need to be reformatted.
This is initially done for the revscoring test image to showcase how we could do this.
Once we decide to proceed it should be applied on the whole repo.
As pre-commit runs on a git repo we do git init and git add before
running the checks. Ideally a CI procedure would clone the git repo
and this step would not be required. This will be the case once
we move this to gitlab.

Bug: T325198
Change-Id: I4aeecc69129a77a24082c9499d5a2e114564a2e9