feat: revscoring kserve upgrade to 0.10
Upgrading the requirements of revscoring models to be able to facilitate both
the upgrade of kserve as well as python 3.9. The reason for bumping kserve to 0.10
is that we need numpy 1.22 that fixes some binary incompatibility issues that
we ran into while using older models.

Also added the headers argument in preprocess and predict functions as it is a
breaking change since kserve 0.10

Bumped the requirements for numpy version in revscoring package to be able to
use numpy above 1.21.

Bug: T325528
Change-Id: I5fd40a0a0de28e075a20a714df1b570547f1e027