Move revscoring model server to fastapi and adapt events
The KServe 0.10 Python release brought a ton of changes,
the most important one is the deprecation of tornado in favor
of the FastAPI specs. We are currently running 0.10 only for
the revscoring model-server (since we need a more up-to-date
version of numpy to be able to run on Debian Bullseye and
Python 3.9), and the events.py module is shared across
model servers.

The solution proposed is a transition step, in which
we move the revscoring's code to be compliant with
FastAPI's spec, and we move events.py to a middle
ground state in which it can be used by all
model servers (with some changes in their code).

While at it, solve an import problem found by Ilias
during testing: the tornado's HTTPError should be
imported with a more specific form to avoid runtime
errors.

Change-Id: Ia8a89d85d04cf123c2bf4afec15758b4b0d74030