test: liftwing manual testing on deployment server
Adding a python script to test the liftwing model servers after deployment.
We got some revision ids for each wiki from the event_sanitized.mediawiki_revision_score table
which are stored in the rev_ids.json file (2 for each wiki).
Also the file deployed_models.yaml has a list of all the deployments available in staging and
production.
The script will query the endpoints and will raise an error if all of the revision ids for
a model server fail to give a 200 response and do not contain the word probability in the text
of the response (since all revscoring models give a response with probabilities in it).
Moreover it checks codfw data center in staging and both eqiad and codfw in production.
The script uses standard python libraries which are installed on deployment server and
has been successully tested on it.
Also a file that creates the yaml files for httpbb has been added. The creation of the httpbb tests
is based on the deployed_models.yaml and rev_ids.json configs

Bug: T327787
Change-Id: I9055fec68be632e47c1db9d6422dcd3de3851f1a