events: fix handling of error responses from eventgate
By default aiohttp doesn't raise an exeception when it gets
a non-200 response code, so the current error handling workflow
doesn't work.

Moreover the aiohttp ClientSession seems to be aimed for
single endpoints, not multiple ones. We already had issues
in the past with this (see T323613) so the easiest solution
in my opinion is to move away from a single shared session
and to use one for each endpoint (in our case, mwapi and
eventgate).

While at it, fix the UserAgent's rendering for revscoring model
servers.

Bug: T325528
Change-Id: Ia16ac3dd5a7b653c72fd3cad23d221ffae399242