revertrisk: create blubberfile and pipeline for each version
Revert-risk model has different versions being developed, so
it would be convenient for each version to have its own
requirement.txt, blubberfile, and pipeline. We have one model
server declaration, since only the import line is different.

So far, we have two versions of the model,
- v1: revertrisk (fully language-agnostic)
- v2: revertrisk-multilingual

In addition, the model name is set by the env var MODEL_NAME
and it'll be the same as the isvc name in the deployment-chart.

Bug: T329936
Change-Id: Ida2c8aba606e433d6c89d645c7cf118a71f6d8f5