feat: Create a migration endpoint between LiftWing/ORES
First version of the API for ores migration. This includes a FastAPI application that has Ores-like endpoints which are
translated to Liftwing API calls. At this points all the endpoints of ORES v3 are implemented but lack support of the model_info field.
For more information check https://ores.wikimedia.org/v3.

For Ores calls that require multiple liftwing calls we perform these requests in an async manner using aiohttp.
Some basic unit testing is also perfomed using pytest-asyncio for the fucntions get_scores (the /v3/scores/{context} endpoint).
More unit tests need to be added, but this patch has been submitted in order to get some early feedback.
Apart from local testing the FastAPI app has been tested on statboxes. Instructions to build the application as well as
how to test it on a statbox are included in README.md file.

Bug: T330414
Change-Id: I57b802d42e761db5226e852b622f003fed956c0e