ores-legacy: add LW error messages and exceptions to response
Add error message to LiftWing responses. As this API is implemented in the context of "do what ORES does"
for each erroneous LW call we include the error message in the final response instead of returning an empty response. Some cases are difficult to cover at the moment :e.g. if wrong model is requested from liftwing with wrong model header the response returned is empty.
Also added some basic unit testing for liftwing requests.

Bug: T330414
Change-Id: I81e7135ee8c8208c499fb8f5f69828cc4112909a