events.py: prioritize the excp handling of ClientResponseError
The code that handles failures in sending events to EventGate
is currently printing generic error messages for ClientError
responses even if the exception type is ClientResponseError.
In the latter case we have a more specific error message,
that should highlight what the problem was logging
the response's body as well.

Bug: T328576
Change-Id: I1c21811934dd68d5d60fa12ac5a1a9ce4d93fbe0