fix(ores-legacy): filter context based on request
The function get_check_models was failing to filter the models' context and as a
result it was returning all the models available regardless of the request. A unit test
has been added to make sure this kind of filtering is working properly.
A liftwing test has been fixed as the wrong kind of error/exception was being expected (result of previous refactor).
Finally, the available_models dict is copied as it was being passed by reference and filtered between calls.

Bug: T330414
Change-Id: Iaca41cc50103c3d0b9433b6479cd458bc1fb19b5