bloom: delay gpu pytorch device context bootstrap
After reading https://github.com/pytorch/pytorch/issues/83973
I am wondering if the errors that we were seeing about
"Cannot re-initialize CUDA in forked subprocess" were due to
having the GPU context initialization in the __init__ constructor.

The idea of the change is to load the model into the GPU's memory
as late as possible but after the initialization of the webserver
(Uvicorn), so that all the processes and fork() calls have
already been made.

Bug: T334583
Change-Id: Icd6406ff3d34b5523588e86938865c0f5c6e9c91