llm: add the ability for facilitate various Open Source LLMs
Add the ability to facilitate multiple open source LLMs using the
same model server. This is needed when different type of model and/or tokenizer is
required. Start with NLLB200 model and extend with others if needed.
Since docker images for LLMs tend to be several GBs in size, the intention is to reuse
the same image and dynamically select a Python Class via configuration, more specifically with an environment
variable.

In this example in order for the NLLB Class to be used one needs to
set the LLM_CLASS env var to "nllb.NLLB"

Bug: T333861
Change-Id: I325d37f9a5e4a5d916a8cf6b163fafbfb318c57e