ores-legacy: avoid any connection pool when calling Lift Wing
The application will run on k8s, and it will use envoy as local
sidecar to contact Lift Wing. We don't need to share the same
ClientSession among multiple requests, and we are investigating
some connection timeouts that may be related to some issues
between the aiohttp's Client session conn pool and envoy's one.

Bug: T341479
Change-Id: I1756e984bcff50f2a0b0db505f7b2d0b8e87d451