events.py: raise debug log to error for eventgate failures
When we need to debug why an event was not posted correctly
to Eventgate, we'd need to find it in the kserve's access logs.

By default we use INFO, so debug is discarded, and we loose
precious info. For example, Changeprop is currently getting
some 500s from kserve outlink and we don't know why.

Change-Id: I8b1fa35563b1bb293be12c3c501fd723975b21c7