python: remove unnecessary self attributes in revscoring's model svc
We are seeing a memory leak over time in revscoring pods, and
we are probably storing/caching memory that we shouldn't.

This change is the first attempt to clear any accidental
reference to objects that may be leaked over time.

The self.extractor reference/attribute is not necessary, since
we reuse it every time. Simply returning it from the API call
should avoid any unnecessary use of self.

Bug: T346445
Change-Id: Iaa4fe9b8c741e5b7d2be9617f8d641296c8eb2e6