revscoring: fix WIKI_URL validation
Convert wiki_url as a class attribute as it is coupled with the current RevscoringModel server.
The main issue this patch solves is an explicit check for WIKI_URL environment variable raising
an error on service spin up.
At the moment if the env var is not specified the server starts correctly but returns the following error when a request is made
which is not at all intuitive
`{"error":"UnboundLocalError : local variable 'resp' referenced before assignment"}`

Change-Id: I9ae9697675a4717192a37c21bb0d75e270e971bb