article-descriptions: enable local run
The model server has been refactored to be able to run locally.
Changes involve:
* rename of article-descriptions directory to article_description in order
to use it as a python module.
* different behaviour for urls included in requests for local vs k8s
deployments. Local runs use external urls.
* Fix headers issue in mwapi.Async requests.
* Cannot load the model locally when low_cpu_mem_usage is True so I have
disabled by default and we can enable it on k8s deployment.

Follow the instructions in the README.md file to run the server locally.

Bug: T351940
Depends-On: I76e343f1b13354bb3f44488bd2f8783d72fae65e
Change-Id: I7f19cd0ba7d49eae678e9f6ad3787a0fc482426f