article-descriptions: update wiki host headers in model-server
To enable the article-descriptions model-server to access both
wikidata.org and wikipedia.org endpoints that are external to
k8s/LiftWing, we have configured the internal host
'api-ro.discovery.wmnet' to use 'wikidata.org' and
'*.wikipedia.org' respectively as the host headers.

The 'api-ro.discovery.wmnet' URL is passed to the model-server
as the WIKI_URL environmental variable using k8s helm configs.

Bug: T343123
Change-Id: I779e5c354968b752f385f7a33c97d589a3363b6c