article-descriptions: add helper function for rest gateway url
Add the following changes as a result of the helper function
* Add get_rest_endpoint_page_summary function that manipulates the
rest gateway endpoint according to the usage (Lift Wing vs local)
* Introduction of the furl package as it allows for easier
url manipulation which we can use in other places as well
* add unit tests for these 2 use cases
* append the top level dir to PYTHONPATH in the docker image in order to
enable access to all modules/directories for import statements. In this case we
need it for the python dir.

A separate patch will be filed to enable unit test runs in ci to minimize the
changes in this patch

Bug: T351940
Change-Id: Id806a7358669338c919c5d7f48ac7068d75aa619