llm: refactor directory structure to treat as python module.
* Refactor the directory structure so that we can treat it as a python module.
The purpose of this is to allow us to have subpackages for submodules
like in nllb where we keep things related to that model e.g. separate readme.md
file, utils etc.
* the class to import is now llm.NLLB instead of nllb.NLLB
* Remove the model-server directory as it doesn't allow us to treat the dir
as a python package (hyphens are not allowed in python) and doesn't serve any
purpose.
* no changes needed in integration/config as the llm/ dir is being monitored
for changes in order to build a new image.

Change-Id: I5bd2181b7e7332f9b2ebf2bcbbffc0fb83729329