article-descriptions: update descartes tag to fix prediction bug
The ML team discovered the cause of a prediction bug in the
article-descriptions model-server and the Research team fixed
it upstream. We have synced the WMF descartes fork with upstream
and created tag 1.0.1. When the new model-server image is built,
it will use the latest WMF descartes repo tagged 1.0.1 that has
the fix.

Bug: T352750
Change-Id: I9c88124f186836c26a80dd61521bb16af0edfd1e