revert-risk: add the option to force HTTP traffic without WIKI_URL set
This change should allow us to test a new Istio configuration, that
in theory should allow us to avoid the use of WIKI_URL (and any
mention to api-ro.discovery.wmnet in our env variable). The idea
is to instruct get_mediawiki_host() to be able to use either HTTP
or HTTPS via environment variable. In production we'll use
FORCE_HTTP=True (to allow the collection of HTTP metrics etc.. by
the istio sidecar) and local testing will default to HTTPS.

Bug: T353622
Change-Id: I57984bc8120e2f9567fcbad167da7865a0ceff16