locust: add article_descriptions load tests
Add a new class for load testing article descriptions model server.
We are using the same inputs as the wrk test.
(except one I removed because it was causing an error).
A test Class for the Cloud VPS instance has been added as well but it is not
enabled by default and can be used on an ad hoc/manual basis as it uses a
different host and using locust with multiple hosts is not straight forward.

Bug: T353952
Change-Id: I0598b97ab20bbfe59b20ea709dc57e7e447033e3