Makefile: maintain models directory structure as analytics repo
The downloaded model(s) were previously added to
'./model_server/models/'. This patch now adds them to
'./models/' and maintains the same directory structure as
the analytics public datasets repo.

This will ensure that different versions of models can be
used by the model-servers.

I also fixed formatting for the prerequisites note in the
README docs.

Bug: T356985
Change-Id: I20fa7fa8a5d4263ddf91636f6aa42250d9fc9dc3