Makefile: add usage steps to RRLA README
Currently, the README file for the RRLA model-server
provides manual instructions on how to set up: a Python venv,
install dependencies, download the model(s), and run the server.

Since we have now automated these steps using a Makefile, this patch
updates the documentation to include both the manual setup instructions
and the new Makefile instructions, giving users the option to choose
which method they prefer.

Bug: T352689
Change-Id: Id2439c5823f7c6e5089ef769c697a46f96537f26