revertrisk-multilingual: reorder requirements.txt
In the previous patch, we encountered a docker image build failure.
The reason might be caused by torch being specified to install the
cpu version in the knowledge integrity.
In this patch, we install the torch-rocm version before installing
knowledge integrity, in hopes of resolving this issue. The image
was successfully built locally, and the size is around 4.6G.

Bug: T356045
Change-Id: I75b1e3943ea0e8eb02e59c32abe2c48103baa05d