ores-legacy: fix mixed boolean and string field
Why:
	The prediction field holds mixed values of either boolean or string.
Pydantic's Union type field also declares a priority meaning that
Union[bool, str] will first try to evaluate a boolean. This results in "1" being
evaluated as True while in some cases like ptwiki-articlequality "1" is
actually a class.
On the other hand models like goodfaith return a boolean in the prediction
field. Although having a mixed schema is not a good practice we want to
maintain this functionality so that we don't break any existing apps.

What:
	To overcome this we change the schema to str and evaluate it as boolean
if the prediction field holds the value of "True", "true", "false" or "False".
We do this using a pydantic validator which runs after the evaluation of the
schema which would allow "1" to stay as is.

Bug: T358953
Change-Id: I6516f5edebf98883205ac5c0b77aa79af452107a