huggingface: add huggingface image
Add a huggingface docker image. This image clones the code from a specific kserve version
and installs the huggingface module which is used to load HF models.

There are two modes of operation as described in the README.md file.
 - If the environment variable MODEL_ID is defined then the server will try to download the model from the
huggingface model repository.
 - If it is empty then the default behavior is to try to load a model that already exists in the /mnt/models directory (this is the default path defined in the kserve code).

At the moment the image is cloning the latest commit of the master branch as new fixes are being
introduced. In the future (with upcoming releases) the production version will be pinned to a specific release.

Bug: T357986
Depends-On: I7236a071a4601d4959d133af92e48b1d1df6c218
Change-Id: I531c549c87aadd51dee8328665d7c4feba33c42a