articletopic-outlink: load model path from environment variable
The current implementation of the model loading method hardcodes
the model path, causing the error in T360177#9634149 when the path
is set via an environment variable. This patch refactors the model
server to accept the model path from an environment variable,
aligning it with other model servers.

This also improves flexibility and ease of configuration,
ensuring smoother integration and deployment in local test environments.

Bug: T360177
Change-Id: I39d742c1cd1ede36862497cc1443aaa5c3969bfb