readability: add entrypoint to set environment variables
After some tests it seems that Pytorch's set_*_threads functions
may not work as expected, and from every github issue that I've
read it is more reliable to just set env variables like
OMP_NUM_THREADS.

This change adds a new entrypoint for readability, that sets
OMP_NUM_THREADS using get_cpu_count(). The benefit of this approach
is that it is easily expandable if needed with more env variables,
and it may be applied to more libraries other than torch.

Bug: T360111
Change-Id: I6ea7a938995533e17665b3df19d184861599d3ac