articletopic-outlink: run model server as python module
In T360177#9636947, the model server failed to run because
the predictor could not import the 'events' module. This patch
refactored the model server to run as python module so that both
the predictor and transformer can easily import the required
modules from 'python/*'.

articletopic-outlink now follows the same pattern as other
model servers which improves maintainability and standardizes
integration with other tools (e.g Makefile)

Bug: T360177
Change-Id: I5135660278550154f2f8f8dfd44960f309dd4bf8