revertrisk: use the Pytorch base image for RRML GPU inference
We changed the base image to a Pytorch base image, which comes with
preinstalled pytorch 2.2.x-rocm. In model.py, if a GPU is detected,
the model will be loaded onto the GPU for inference.

The new image size is 13.6GB, and you can find the size of layers
at https://phabricator.wikimedia.org/T356045#9700139

Bug: T356045
Change-Id: I40d01e1b5deceddccd1d9d6e1ca5a567c10a33a9