utils: slow function execution wrapper
Add a decorator that can be used both for async and sync functions that will log
a warning message if it passes the threshold (a threshold can be passed as an
argument to the decorator).

The purpose is to use this decorator so that we can debug functions that take too
much time to execute.

Bug: T362663
Change-Id: Iff435042c2c73fca5cf8905fa9ce9e94ab565403