revscoring_model: inspect mw-api-cache for MP preprocess
If the size of the MW-API rev-id content JSON payload returned
is more than a certain threshold (by default, 100KB), then
it is "safe" to offload it to a separate process (if PREPROCESS_MP
is enabled) because the penalty for serialization/deserialization
will likely be not impacting to the overall latency.
This new workflow should allow heavy rev-id contents to be processed
separately from the main ioloop thread, avoid blocking other
unrelated calls (as much as possible).

Bug: T363336
Change-Id: I179f06a7715d43b364b99dec7e1c8fc69443fd7b