articlequality: add feature preprocess
- Add a function called load_quality_max_featurevalues() to read
the table of maximum feature values used for feature normalization.
There is one table in the data/ directory, which is used in the
notebook. It contains data from 6 wikis and includes new features
(max_srcs, infobox, and mbox).

- Functions such as get_article_features and normalize_features, which
come from the notebook, remain unchanged.

- get_article_html from the notebook is used to obtain an article
revision’s HTML. This should be modified to an asynchronous function
using the aiohttp session.

Bug: T360455
Change-Id: I36e854b6318748bf7018a4e2eda17c7e296093c7