test: update articletopic-transformer unit test
In T384993#10502948, this transformer's get_outlinks test case is
failing because new outlinks were added to the Muumi article on fiwiki.

We have verified that these new outlinks exist in this article as shown
in: T384993#10502966

This patch updates the test case with wikibase ids associated to the
new outlinks in this article.

Bug: T384993
Change-Id: Ib047cd1b1de5da11fc747574b860c5882ce1c18d