edit-check locust tests: detect and report failed instances in batch responses
The previous tests (until now) were reporting no failures even if some of the requests were containing malformed instances. This is due to the architecture of the current service for the time being. The service returns a 200 for the batch request, and individual status codes in the response for each of the instances.
 - Parse the responses from the locust tests
 - Report errors
 - Run only valid requests for locust

Bug: T392460
Change-Id: I75357a937f40ae883617b09cff5c2fcbd89a9d5d