outlink-topic-model: Merge transformer and predictor pods.
We combine the transformer and predictor pods into a single pod to
simplify the setup, lower the communication overhead between the pods
and make it easier to integrate Caching mechanism, which will be
introduced in the upcoming task T401778.
In this patch, we move all processing logic to the original predictor
pod, but  still keep the transformer code until we finish testing
the single-pod deployment.

Bug: T404294
Change-Id: Iddfcf9d9dcbec0d655434228f54f3d11b7083653